pyproject.toml
[project]
name = "boomi-mcp-server"
version = "0.1.0"
description = "FastMCP server for Boomi API integration with secure credential storage"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
